Update#2

+++ RSPiX/Src/GREEN/3D/pipeline.cpp
@@ -107,9 +107,10 @@ int16_t RPipeLine::CreateShadow(int16_t sAngleY,
 	// Create the shadow transform:
 	m_tShadow.Make1();
 	m_dShadowScale = dTanDeclension;
-	m_tShadow.T[1 + ROW0] = m_dShadowScale * rspCos(sAngleY);
-	m_tShadow.T[1 + ROW1] = 0.0;
-	m_tShadow.T[1 + ROW2] = m_dShadowScale * rspSin(sAngleY);
+	m_tShadow.T[1 + ROW0] = static_cast<float>(m_dShadowScale * rspCos(sAngleY));
+	m_tShadow.T[1 + ROW1] = 0.0f;
+	m_tShadow.T[1 + ROW2] = static_cast<float>(m_dShadowScale * rspSin(sAngleY));
+
 
 	// Allocate the buffer, if applicable:
 	if (sBufSize <= 0)	// default case:
@@ -428,8 +429,9 @@ void RPipeLine::BoundingSphereToScreen(RP3d& ptCenter, RP3d& ptRadius,
 	m_sY = m_sCenY - sScreenRadius;
 	m_sZ = m_sCenZ - sScreenRadius;
 
-	m_sW = m_sH = (sScreenRadius * 2.);
-	m_sD = (sScreenRadius * 2.);
+	m_sW = m_sH = static_cast<int16_t>(sScreenRadius * 2.0);
+	m_sD = static_cast<int16_t>(sScreenRadius * 2.0);
+
 
 	m_sUseBoundingRect = TRUE;
 	}
